now creating URI with netaporter.uri
